requirements.txt
fastapi
uvicorn     
python-decouple       
openai<1.0
requests
python-multipart      
